feat/qwen3_5_fuse_shared_expert

See merge request mcc-ai-serving/sglang!331

+++ python/sglang/srt/hardware_backend/musa/jit_kernel/tilelang/_atomic_helper.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdint.h>
+
 #include <tl_templates/musa/cvt.h>
 
 __device__ __forceinline__ int sgl_tl_atomic_add_offset(int *base, int offset,
@@ -8,7 +10,7 @@ __device__ __forceinline__ int sgl_tl_atomic_add_offset(int *base, int offset,
 }
 
 __device__ __forceinline__ void sgl_tl_store_fp8e4m3x4(fp8_e4_t *base,
-                                                       int offset, float x0,
+                                                       int64_t offset, float x0,
                                                        float x1, float x2,
                                                        float x3) {
   const float4 values = {x0, x1, x2, x3};
@@ -19,16 +21,16 @@ __device__ __forceinline__ void sgl_tl_store_fp8e4m3x4(fp8_e4_t *base,
 
 __device__ __forceinline__ void sgl_tl_copy_bf16x8(bfloat16_t *dst,
                                                    const bfloat16_t *src,
-                                                   int dst_offset,
-                                                   int src_offset) {
+                                                   int64_t dst_offset,
+                                                   int64_t src_offset) {
   const int4 value = *reinterpret_cast<const int4 *>(src + src_offset);
   *reinterpret_cast<int4 *>(dst + dst_offset) = value;
 }
 
 __device__ __forceinline__ void sgl_tl_copy_fp8x16(fp8_e4_t *dst,
                                                    const fp8_e4_t *src,
-                                                   int dst_offset,
-                                                   int src_offset) {
+                                                   int64_t dst_offset,
+                                                   int64_t src_offset) {
   const int4 value = *reinterpret_cast<const int4 *>(src + src_offset);
   *reinterpret_cast<int4 *>(dst + dst_offset) = value;
 }
